refactoring, removed unnecessary code and fixed imports
